Rename HydroOrb.py to Hydrogen_Orbitals..py
The objective of this program is to plot the non relativistic wave function of the hydrogen atom. For that, the program solves the the partial differential equation using separation of variables. Then it plot te eigenstates of the normalized wavefunction.
The input are the Principal (n), Azimuthal (l) and Magnetic (m) Quantum Numbers. The three numbers must be integers, following these rules:
n = 1, 2, 3, ...
l = 0, 1, 2, ..., n-1
m = -m, ..., 0, ... m
The result cute plot of the orbitals of the atom with those quantum numbers.